add an example script
I had a difficult time parsing out the example from the synopsis and I
had to dig into the tests. This commit adds a simple eg/ directory with
the script that I ended up writing. It uses NetAddr::IP to take a list
of CIDR blocks and ping all of the hosts.